third_party/blink/web_tests/flag-specific/highdpi/fast/compositor-wheel-scroll-latching/animated-scroll/touchpad-scroll-impl-to-main-expected.txt

This is a testharness.js-based test.
FAIL Scroller should continue to scroll when MT takes over scroll handling. assert_equals: div must be fully scrolled expected 315 but got 314.6666564941406
Harness: the test ran to completion.
